modificaciones en tp2

+++ tp2/cms.h
@@ -9,7 +9,7 @@ typedef struct cms cms_t;
 
 /* Crea el cms
  */
-cms_t* cms_crear(size_t tam);
+cms_t* cms_crear(int tam);
 
 /* Incrementa en uno la cantidad de apariciones de la clave,
  * Pre: La estructura cms fue inicializada
@@ -20,7 +20,7 @@ void cms_guardar(cms_t* cms, const char* clave);
 /* Devuelve la cantidad de apariciones de la clave (aproximada)
  * Pre: La estructura cms fue inicializada
  */
-size_t cms_obtener(const cms_t *cms, const char *clave);
+int cms_obtener(const cms_t *cms, const char *clave);
 
 /* Destruye la estructura liberando la memoria pedida
  * Pre: La estructura cms fue inicializada
